+++ CHANGELOG.md
@@ -8,5 +8,6 @@
 * Add option post_type
 * Add screenshots
 * Initial code release 
+* Remove featured image meta data
 * Remove unused CSS stylesheet
 * Update description